# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Template Simbi API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/loans/create": {
            "post": {
                "tags": [
                    "Loans"
                ],
                "summary": "Adiciona um novo emprestimo na API",
                "description": "Permite adicionar um novo emprestimo",
                "operationId": "fa00e38f31f936e4936eb74b0a9e8df6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "firstName",
                                    "lastName"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "cpf": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "exit_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "last_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "authorId": {
                                        "type": "string"
                                    },
                                    "bookId": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Loan Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "13b35be6-65a7-4d7f-9ad2-8caaf3c75344"
                                        },
                                        "first_name": {
                                            "type": "string",
                                            "example": "Primeiro nome"
                                        },
                                        "last_name": {
                                            "type": "string",
                                            "example": "Ultimo nome"
                                        },
                                        "cpf": {
                                            "type": "string",
                                            "example": "000.000.000-00"
                                        },
                                        "phone": {
                                            "type": "string",
                                            "example": "(00)0000-0000"
                                        },
                                        "address": {
                                            "type": "string",
                                            "example": "Endereço do cliente à realizar o emprestimo"
                                        },
                                        "lastName": {
                                            "type": "string",
                                            "example": "Ultimo nome"
                                        },
                                        "exitDate": {
                                            "type": "string",
                                            "example": "2022-12-14T22:24:26+00:00"
                                        },
                                        "deliveryDate": {
                                            "type": "string",
                                            "example": "2022-12-14T22:24:26+00:00"
                                        },
                                        "author": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "13b35be6-65a7-4d7f-9ad2-8caaf3c75344"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Autor Y"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "book": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "13b35be6-65a7-4d7f-9ad2-8caaf3c75344"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Book Z"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "createdAt": {
                                            "type": "string",
                                            "example": "2022-12-14T22:24:26+00:00"
                                        },
                                        "updatedAt": {
                                            "type": "string",
                                            "example": "2022-12-14T22:24:26+00:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requisição com erro",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Bad Request"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Proibido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Forbidden Error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthorized Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Internal server error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}